Gaia file source.
