+++ src/org/aavso/tools/vstar/ui/resources/RevisionAccessor.java
@@ -17,14 +17,17 @@
  */
 package org.aavso.tools.vstar.ui.resources;
 
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+
 /**
  * The purpose of this class is to provide access to subversion revision number.
  */
 
 public class RevisionAccessor {
 
-	private static String REVISION = "6d11427";
-	private static String BUILD_TIME = "2022-11-14 23:12";
+	private static String REVISION = "15922a20";
+	private static String BUILD_TIME = "2023-03-24 18:26";
 
 	/**
 	 * Get the latest git revision
